hw: intc: make apic configurable
If the irq offload to kernel side, the apic in qemu is
not usefull and can get compiled out.
