+++ default-configs/x86_64-softmmu.mak
@@ -49,6 +49,7 @@ CONFIG_SGA=y
 CONFIG_LPC_ICH9=y
 CONFIG_PCI_Q35=y
 CONFIG_APIC=y
+CONFIG_APIC_COMMON=y
 CONFIG_IOAPIC=y
 CONFIG_IOAPIC_COMMON=y
 CONFIG_PVPANIC=y
