+++ default-configs/x86_64_virt-softmmu.mak
@@ -20,7 +20,7 @@ CONFIG_VIRTIO_PCI=y
 
 CONFIG_ACPI=y
 CONFIG_ACPI_X86=y
-CONFIG_APIC=y
+CONFIG_APIC_COMMON=y
 CONFIG_IOAPIC_COMMON=y
 CONFIG_NVDIMM=y
 CONFIG_ACPI_NVDIMM=y
