+++ hw/intc/Makefile.objs
@@ -22,11 +22,12 @@ common-obj-$(CONFIG_ARM_GIC) += arm_gicv3_its_common.o
 common-obj-$(CONFIG_OPENPIC) += openpic.o
 common-obj-y += intc.o
 
-ifeq ($(CONFIG_APIC),y)
-common-obj-$(call lnot,$(CONFIG_I8259)) += stub.o
+ifeq ($(CONFIG_APIC_COMMON),y)
+common-obj-$(call lnot,$(call lor,$(CONFIG_I8259),$(CONFIG_APIC))) += stub.o
 endif
 
-obj-$(CONFIG_APIC) += apic.o apic_common.o
+obj-$(CONFIG_APIC) += apic.o
+obj-$(CONFIG_APIC_COMMON) += apic_common.o
 obj-$(CONFIG_ARM_GIC_KVM) += arm_gic_kvm.o
 obj-$(call land,$(CONFIG_ARM_GIC_KVM),$(TARGET_AARCH64)) += arm_gicv3_kvm.o
 obj-$(call land,$(CONFIG_ARM_GIC_KVM),$(TARGET_AARCH64)) += arm_gicv3_its_kvm.o
